Refactor errors and constants into separate file
