+++ compliant-reward-distribution/indexer-and-server/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod constants;
 pub mod db;
+pub mod error;
 pub mod types;
 pub use crate::db::DatabasePool;
